+++ src/client/js/global.js
@@ -2,6 +2,6 @@
  * pwix:tolert/src/client/js/global.js
  */
 
-tlTolert = {
+Tolert = {
     _conf: {}
 };