+++ src/client/js/private_fn.js
@@ -2,17 +2,17 @@
  * pwix:tolert/src/client/js/private_fn.js
  */
 
-tlTolert._client = {
+Tolert._client = {
 
     deleteOldest(){
-        const array = tlTolert._client.Messages.find({}, { sort: { stamp: 1 }}).fetch();
+        const array = Tolert._client.Messages.find({}, { sort: { stamp: 1 }}).fetch();
         if( array.length ){
             const height = parseInt( $( '.tl-display .alert[data-tl-count="0"]' ).css( 'height' ));
-            const start = tlTolert._conf.stackFromTop ? 'top' : 'bottom';
+            const start = Tolert._conf.stackFromTop ? 'top' : 'bottom';
             let args = {};
             args[start] = '-=' + (height+CSS_MARGIN) + 'px'
             $( '.tlMessage .alert' ).css( args );
-            tlTolert._client.Messages.remove({ _id: array[0]._id });
+            Tolert._client.Messages.remove({ _id: array[0]._id });
         }
     },
 
@@ -21,7 +21,7 @@ tlTolert._client = {
     //  - msg: the message to be displayed
     // add a stamp to be able to identify the oldest later
     display( o ){
-        tlTolert._client.Messages.insert({ ...o, stamp: Date.now() });
+        Tolert._client.Messages.insert({ ...o, stamp: Date.now() });
     },
 
     // the collection itself